Added App component with global theme
