+++ src/Portfolio/Components/About/index.js
@@ -15,7 +15,7 @@ const useStyles = makeStyles((theme) => ({
   aboutPaper: {
     borderRadius: 0,
     height: '100%',
-    background: theme.palette.background.circularGradient,
+    backgroundColor: 'transparent',
     overflowY: 'scroll',
   },
 }));
