+++ src/Portfolio/Components/Projects/index.js
@@ -17,7 +17,7 @@ const useStyles = makeStyles((theme) => ({
   projectsPaper: {
     borderRadius: 0,
     height: '100%',
-    background: theme.palette.background.circularGradient,
+    backgroundColor: 'transparent',
     overflowY: 'scroll',
   },
 }));
